requirements/base.txt
Django==1.8.2
django-braces==1.8.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-floppyforms==1.4.1
django-jsonfield==0.9.13
django-toolbelt==0.0.1
vld==0.0.1
